lama_clean/requirements.txt
# LaMa Requirements
# Note: PyTorch will be installed automatically by setup.py based on your CUDA version
# For manual installation, see README.md

# Core dependencies
numpy>=1.19.2
opencv-python>=4.5.0  
Pillow>=8.0.0
scikit-image>=0.17.2
scipy>=1.5.4
tqdm>=4.60.0

# Deep Learning
# torch>=1.12.0  # Installed by setup.py
# torchvision>=0.13.0  # Installed by setup.py
pytorch-lightning>=1.6.0,<2.0.0
kornia>=0.5.0
einops>=0.4.1

# Configuration and Logging
hydra-core>=1.1.0
omegaconf>=2.1.0
PyYAML>=5.4.0
tensorboard>=2.7.0

# Data Processing
albumentations>=0.5.2
webdataset>=0.1.40

# Visualization
matplotlib>=3.3.4
pandas>=1.1.5

# Model Export
onnx>=1.10.0

# Optional: Install with pip install -e ".[gpu]" for GPU support
# onnxruntime-gpu>=1.10.0  # For GPU
# onnxruntime>=1.10.0  # For CPU

# Optional: Install with pip install -e ".[export]" for advanced export features  
# onnx-simplifier>=0.3.6
# onnxconverter-common>=1.8.0